Added user action to player comparison screen
Result of user choice is displayed and the score is incremented if correct

+++ FanDuelTest/PlayerObject.h
@@ -14,13 +14,13 @@
 -(instancetype)initWithId:(int)player_id
                First_Name:(NSString *)first_name
                 Last_Name:(NSString *)last_name
-                     FPPG:(int)fppg
+                     FPPG:(float)fppg
                 ImageDict:(NSDictionary *)imageDict;
 
 @property (nonatomic) int player_id;
 @property (nonatomic, strong) NSString *first_name;
 @property (nonatomic, strong) NSString *last_name;
-@property (nonatomic) int fppg;
+@property (nonatomic) float fppg;
 @property (nonatomic, strong) NSDictionary *imageDict;
 
 @property (nonatomic, strong) NSString *defaultImageURL;
